requirements.txt
aiohttp
aiohttp-devtools
pandas
plotly
colorama

black
mypy
pytest
pytest-factoryboy
